requirements.txt
pandas
dash
dash-bootstrap-components
snowflake-connector-python
snowflake-connector-python[pandas]
SQLAlchemy
python-dotenv